refactor: 테스트에 구현된 클래스들을 main 패키지로 이동
