+++ README.md
@@ -23,7 +23,7 @@
   - [x] 숫자는 반드시 6개여야 한다
   - [x] 숫자는 1~45까지의 숫자여야 한다
 - 구매한 로또와 당첨 번호 비교
-  - [] 구매한 로또 목록과 당첨 번호를 비교한다
-  - [] 등수 별로 개수를 기록한다
+  - [x] 구매한 로또 목록과 당첨 번호를 비교한다
+  - [x] 등수 별로 개수를 기록한다
 - 로또 통계 
   - [] 구매한 금액 대비 당첨 금액으로 수익률을 계산한다
