+++ src/main/kotlin/lotto/LottoResults.kt
@@ -1,0 +1,5 @@
+package lotto
+
+class LottoResults(ranks: List<LottoRank>) {
+    val resultCountByRank: Map<LottoRank, Int> = ranks.groupingBy { it }.eachCount()
+}
