+++ src/test/kotlin/lotto/LottoResultsCheckerTest.kt
@@ -3,41 +3,6 @@ package lotto
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
 
-class LottoResultChecker {
-    fun check(lottos: List<Lotto>, winningNumbers: WinningNumbers): LottoResults {
-        val matchCounts = lottos.map { lotto ->
-            lotto.calculateMatchCount(winningNumbers)
-        }
-
-        val ranks = matchCounts.filter { matchCount -> LottoRank.isInTheRank(matchCount) }
-            .map { matchCount -> LottoRank.from(matchCount) }
-
-        return LottoResults(ranks)
-    }
-
-}
-
-class LottoResults(ranks: List<LottoRank>) {
-    val resultCountByRank: Map<LottoRank, Int> = ranks.groupingBy { it }.eachCount()
-}
-
-enum class LottoRank(private val matchCount: Int, private val prizeAmount: Long) {
-    FIRST(6, 2_000_000_000),
-    SECOND(5, 1_500_000),
-    THIRD(4, 50_000),
-    FORTH(3, 5_000),;
-
-    companion object {
-        fun from(matchCount: Int): LottoRank {
-            return entries.first { rank -> rank.matchCount == matchCount }
-        }
-
-        fun isInTheRank(matchCount: Int): Boolean {
-            return matchCount >= FORTH.matchCount
-        }
-    }
-}
-
 class LottoResultsCheckerTest {
     @Test
     fun `구매한 로또 목록과 당첨 번호를 비교해서 등수 별 개수를 기록한다`() {
